Go back to Qt6, it's just easier

+++ src/libqcontiguouscache.h
@@ -21,7 +21,7 @@ extern "C" {
 typedef struct QContiguousCacheData QContiguousCacheData;
 #endif
 
-QContiguousCacheData* QContiguousCacheData_AllocateData(int size, int alignment);
+QContiguousCacheData* QContiguousCacheData_AllocateData(ptrdiff_t size, ptrdiff_t alignment);
 void QContiguousCacheData_FreeData(QContiguousCacheData* data);
 void QContiguousCacheData_Delete(QContiguousCacheData* self);
